[apps] Split on_boarding/language_controller in two
