+++ apps/on_boarding/language_controller.cpp
@@ -5,35 +5,21 @@
 namespace OnBoarding {
 
 LanguageController::LanguageController(Responder * parentResponder, LogoController * logoController, UpdateController * updateController) :
-  ViewController(parentResponder),
-  m_logoController(logoController),
+  Shared::LanguageController(parentResponder),
 #if OS_WITH_SOFTWARE_UPDATE_PROMPT
   m_updateController(updateController),
 #endif
-  m_selectableTableView(this, this, 0, 1, (Ion::Display::Height - I18n::NumberOfLanguages*Metric::ParameterCellHeight)/2, Metric::CommonRightMargin, 0, Metric::CommonLeftMargin, this)
+  m_logoController(logoController)
 {
-  for (int i = 0; i < I18n::NumberOfLanguages; i++) {
-    m_cells[i].setMessageFontSize(KDText::FontSize::Large);
-  }
-}
-
-View * LanguageController::view() {
-  return &m_selectableTableView;
 }
 
 void LanguageController::reinitOnBoarding() {
-  m_selectableTableView.deselectTable();
-  selectCellAtLocation(0, 0);
+  resetSelection();
   app()->displayModalViewController(m_logoController, 0.5f, 0.5f);
 }
 
-void LanguageController::didBecomeFirstResponder() {
-  app()->setFirstResponder(&m_selectableTableView);
-}
-
 bool LanguageController::handleEvent(Ion::Events::Event event) {
-  if (event == Ion::Events::OK || event == Ion::Events::EXE) {
-    GlobalPreferences::sharedGlobalPreferences()->setLanguage((I18n::Language)(selectedRow()+1));
+  if (Shared::LanguageController::handleEvent(event)) {
 #if OS_WITH_SOFTWARE_UPDATE_PROMPT
     app()->displayModalViewController(m_updateController, 0.5f, 0.5f);
 #else
@@ -47,28 +33,6 @@ bool LanguageController::handleEvent(Ion::Events::Event event) {
     return true;
   }
   return false;
-}
-
-int LanguageController::numberOfRows() {
-  return I18n::NumberOfLanguages;
-}
-
-KDCoordinate LanguageController::cellHeight() {
-  return Metric::ParameterCellHeight;
-}
-
-HighlightCell * LanguageController::reusableCell(int index) {
-  return &m_cells[index];
-}
-
-int LanguageController::reusableCellCount() {
-  return I18n::NumberOfLanguages;
-}
-
-void LanguageController::willDisplayCellForIndex(HighlightCell * cell, int index) {
-  MessageTableCell * myCell = (MessageTableCell *) cell;
-  I18n::Message languages[I18n::NumberOfLanguages] = {I18n::Message::English, I18n::Message::French, I18n::Message::Spanish, I18n::Message::German, I18n::Message::Portuguese};
-  myCell->setMessage(languages[index]);
 }
 
 }
