+++ apps/shared/Makefile
@@ -24,6 +24,7 @@ app_objs += $(addprefix apps/shared/,\
   interactive_curve_view_range_delegate.o\
   interval.o\
   interval_parameter_controller.o\
+  language_controller.o\
   list_controller.o\
   list_parameter_controller.o\
   memoized_curve_view_range.o\
